feat:Emailクラス削除とMembershipStatusのインポートを止める
