--- backend/src/domain/entity/participant/email.ts
@@ -1,19 +1,0 @@
-export class Email {
-  private readonly value: string
-
-  constructor(email: string) {
-    if (!this.isValid(email)) {
-      throw new Error('Invalid email address provided.')
-    }
-    this.value = email
-  }
-
-  private isValid(email: string): boolean {
-    const regex = /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/
-    return regex.test(email)
-  }
-
-  public toString() {
-    return this.value
-  }
-}
