--- backend/src/domain/entity/participant/membershipStatus.ts
@@ -1,1 +1,0 @@
-export type MembershipStatus = 'Active' | 'Inactive' | 'Left'
